avltree: update that probably won't make a difference
from:
https://github.com/greensky00/avltree/commit/0dbd2fb4ce26b3da38394baeb9f9bd4899879824

message:
Fix compiler issues
* Replaced `__inline` with `static inline`.
* Replaced `uint64_t` with `uintptr_t` for 32-bit platforms.

The valgrind problems show on linux-x86_64 so I doubt this improves anything.

+++ external/avltree/avltree.h
@@ -60,12 +60,12 @@ struct avl_tree{
 #endif
 
 #define avl_parent(node) \
-        ((struct avl_node *)((uint64_t)(node)->parent & ~0x3))
+        ((struct avl_node *)((uintptr_t)(node)->parent & ~0x3))
 
 #ifdef _AVL_SEPARATE_PARENT_BF
     #define avl_bf(node) ((node)->bf)
 #else
-    #define avl_bf(node) (((int)((uint64_t)(node)->parent & 0x3)) - 1)
+    #define avl_bf(node) (((int)((uintptr_t)(node)->parent & 0x3)) - 1)
 #endif
 
 // *a < *b : return neg
